refactor code to call `fetchBasicGasAndTimeEstimates` method once
