+++ ui/app/components/app/transaction-time-remaining/transaction-time-remaining.component.js
@@ -17,26 +17,10 @@ export default class TransactionTimeRemaining extends PureComponent {
     statusKey: PropTypes.string,
     className: PropTypes.string,
     currentTimeEstimate: PropTypes.string,
-    fetchBasicGasAndTimeEstimates: PropTypes.func,
-    fetchGasEstimates: PropTypes.func,
-    blockTime: PropTypes.oneOfType([
-      PropTypes.string,
-      PropTypes.number,
-    ]),
   }
 
   render () {
     const { className, statusKey } = this.props
-
-    if (statusToTextHash[statusKey] === 'pending') {
-      const promise = this.props.fetchBasicGasAndTimeEstimates()
-        .then(basicEstimates => basicEstimates.blockTime)
-
-      promise
-        .then(blockTime => {
-          this.props.fetchGasEstimates(blockTime)
-        })
-    }
 
     return (
       <div className={className}>
